refactor: - Refactor delete_profile_menu() to return Result - Properly handle Result in option_control() function - Add error propagation using ? operator - Update related functions to handle potential errors - Enhance user feedback for error cases

+++ src/main.rs
@@ -15,7 +15,7 @@ fn main() -> Result<(), Box<dyn std::error::Error>> {
     let mut stdout = stdout();
 
     utils::common::clear_terminal();
-    let mut selected_option: u8 = 1;
+    let mut selected_option: usize = 1;
 
     //thread::sleep(Duration::from_secs(2));
